# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GMI Corporate Website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Business {
  id: number;
  name: string;
  slug: string;
  industry: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  colorAccent?: string | null;
  featured?: boolean;
  order: number;
  services?: string[];
  /** @nullable */
  targetAudience?: string | null;
  /** @nullable */
  website?: string | null;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  categorySlug?: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  /** @nullable */
  businessSlug?: string | null;
  tags?: string[];
}

export interface ProductList {
  items: Product[];
  total: number;
}

export interface ProductCategory {
  slug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  count: number;
}

export interface NewsArticle {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  authorName?: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  publishedAt: string;
  tags?: string[];
}

export interface NewsList {
  items: NewsArticle[];
  total: number;
}

export interface Job {
  id: number;
  title: string;
  department: string;
  location: string;
  type: string;
  /** @nullable */
  description?: string | null;
  requirements?: string[];
  postedAt: string;
  status?: string;
}

export interface ApplicationInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  phone: string;
  positionApplying: string;
  linkedinUrl?: string;
  coverLetter?: string;
  /** @nullable */
  jobId?: number | null;
}

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  subject: string;
  /** @minLength 10 */
  message: string;
}

export type BusinessInquiryInputInquiryType = typeof BusinessInquiryInputInquiryType[keyof typeof BusinessInquiryInputInquiryType];


export const BusinessInquiryInputInquiryType = {
  distributor: 'distributor',
  investor: 'investor',
  partnership: 'partnership',
  export: 'export',
  other: 'other',
} as const;

export interface BusinessInquiryInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone: string;
  company?: string;
  inquiryType: BusinessInquiryInputInquiryType;
  /** @minLength 10 */
  message: string;
  region?: string;
}

export interface SubmitResponse {
  success: boolean;
  message: string;
}

export interface CompanyStats {
  subsidiaries: number;
  products: number;
  districtsCovered: number;
  yearsActive: number;
  dealerCount: number;
  countriesExported: number;
}

export interface Crop {
  id: number;
  name: string;
  /** @nullable */
  englishName?: string | null;
  category: string;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  soilType?: string | null;
  /** @nullable */
  fertilizerNotes?: string | null;
  /** @nullable */
  irrigationNotes?: string | null;
  /** @nullable */
  seedVarietyRef?: string | null;
  /** @nullable */
  expectedYield?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CropList {
  items: Crop[];
  total: number;
}

export interface Disease {
  id: number;
  name: string;
  /** @nullable */
  cropId?: number | null;
  /** @nullable */
  cropName?: string | null;
  category: string;
  symptoms?: string[];
  /** @nullable */
  causeType?: string | null;
  /** @nullable */
  causeNotes?: string | null;
  treatmentText: string;
  preventionSteps?: string[];
  relatedProductIds?: number[];
  /** @nullable */
  imageUrl?: string | null;
}

export interface DiseaseList {
  items: Disease[];
  total: number;
}

export interface Season {
  id: number;
  name: string;
  /** @nullable */
  englishName?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  months?: string | null;
  /** @nullable */
  sowingWindow?: string | null;
  /** @nullable */
  transplantingWindow?: string | null;
  /** @nullable */
  harvestWindow?: string | null;
  applicableCrops?: string[];
}

export type WeatherResponseCurrent = {
  temp?: number;
  feelsLike?: number;
  humidity?: number;
  windSpeed?: number;
  description?: string;
  icon?: string;
};

export type WeatherResponseForecastItem = {
  dt?: number;
  temp?: number;
  description?: string;
  icon?: string;
};

export interface WeatherResponse {
  district: string;
  current?: WeatherResponseCurrent;
  forecast?: WeatherResponseForecastItem[];
  fetchedAt?: string;
  source?: string;
}

export interface AdvisoryInquiryInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 5 */
  phone: string;
  district?: string;
  crop?: string;
  /** @minLength 10 */
  question: string;
}

export type ListBusinessesParams = {
featured?: boolean;
};

export type ListProductsParams = {
category?: string;
featured?: boolean;
limit?: number;
offset?: number;
};

export type ListNewsParams = {
category?: string;
limit?: number;
offset?: number;
};

export type ListJobsParams = {
department?: string;
};

export type ListCropsParams = {
category?: string;
season?: string;
search?: string;
limit?: number;
offset?: number;
};

export type ListDiseasesParams = {
category?: string;
cropId?: number;
search?: string;
limit?: number;
offset?: number;
};

export type GetWeatherParams = {
district?: string;
};