# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GMI Corporate Website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all GMI subsidiary businesses
      parameters:
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
  /businesses/{slug}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a single business by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Business detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products with optional category filter
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of news articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsList"
  /news/{slug}:
    get:
      operationId: getNewsArticle
      tags: [news]
      summary: Get a single news article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
...
components:
  schemas:
...
    Business:
      type: object
      required: [id, name, slug, industry, description, order]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        industry:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        colorAccent:
          type: ["string", "null"]
        featured:
          type: boolean
        order:
          type: integer
        services:
          type: array
          items:
            type: string
        targetAudience:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
    Product:
      type: object
      required: [id, name, category, description]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        categorySlug:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        businessSlug:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    ProductList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
...
    NewsArticle:
      type: object
      required: [id, title, slug, content, category, publishedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        publishedAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
    NewsList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
        total:
          type: integer
...